Update main.py
Changed int to float

+++ getting-started-with-python-main/error-handling-with-try-and-except/main.py
@@ -5,18 +5,18 @@ user_input2 = input("Input your second number: ")
 is_good = True
 
 try:
-    int(user_input1)
+    float(user_input1)
 except:
     print("Your first number needs to be an integer!")
     is_good = False
 try:
-    int(user_input2)
+    float(user_input2)
 except:
     print("Your second number needs to be an integer!")
     is_good = False
 
 if (is_good):
-   sum = int(user_input1) + int(user_input2)    
+   sum = float(user_input1) + float(user_input2)    
    print("Your sum is :", sum)
 else:
    print ()
